refactor: move `ready` command from `Changelog` to `Version` namespace
This command is about marking a specific **version** as ready in the
changelog, so it belongs in the `Version` namespace.

As such, the command changes to `version:ready`.
