+++ bin/keep-a-changelog
@@ -53,9 +53,9 @@ $application->addCommands([
     new Entry\EntryCommand($dispatcher, 'entry:removed'),
     new Entry\EntryCommand($dispatcher, 'entry:fixed'),
     new Changelog\NewCommand($dispatcher, 'changelog:new'),
-    new Changelog\ReadyCommand($dispatcher, 'changelog:ready'),
     new Version\EditCommand($dispatcher, 'version:edit'),
     new Version\ListCommand($dispatcher, 'version:list'),
+    new Version\ReadyCommand($dispatcher, 'version:ready'),
     new Version\ReleaseCommand($dispatcher, 'version:release'),
     new Version\RemoveCommand($dispatcher, 'version:remove'),
     new Version\ShowCommand($dispatcher, 'version:show'),
