+++ test/Changelog/ReadyLatestChangelogEventTest.php
@@ -7,9 +7,9 @@
 
 declare(strict_types=1);
 
-namespace PhlyTest\KeepAChangelog\Changelog;
+namespace PhlyTest\KeepAChangelog\Version;
 
-use Phly\KeepAChangelog\Changelog\ReadyLatestChangelogEvent;
+use Phly\KeepAChangelog\Version\ReadyLatestChangelogEvent;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Psr\EventDispatcher\EventDispatcherInterface;
